adapt tests
