+++ app/src/androidTest/java/com/github/sdp/ratemyepfl/fragment/review/AddReviewFragmentTest.kt
@@ -5,20 +5,18 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.closeSoftKeyboard
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.action.ViewActions.typeText
+import androidx.test.espresso.action.ViewActions.*
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
-import com.github.sdp.ratemyepfl.R
 import com.github.sdp.ratemyepfl.activity.ReviewActivity
 import com.github.sdp.ratemyepfl.auth.FakeConnectedUser
-import com.github.sdp.ratemyepfl.database.reviewable.ClassroomRepository
 import com.github.sdp.ratemyepfl.model.items.Classroom
-import com.github.sdp.ratemyepfl.model.items.Restaurant
 import com.github.sdp.ratemyepfl.model.review.ReviewRating
 import com.github.sdp.ratemyepfl.model.serializer.putExtra
 import com.github.sdp.ratemyepfl.utils.CustomViewActions
+import com.github.sdp.ratemyepfl.R
 import com.github.sdp.ratemyepfl.utils.CustomViewActions.RatingAction.performSetRating
+import com.github.sdp.ratemyepfl.utils.CustomViewActions.ViewPagerAction
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.After
@@ -38,12 +36,12 @@ class AddReviewFragmentTest {
         val reviewable = Classroom("Fake id", 0.0, 0)
         FakeConnectedUser.instance = FakeConnectedUser.Instance.FAKE_USER_1
         val intent = Intent(ApplicationProvider.getApplicationContext(), ReviewActivity::class.java)
-        intent.putExtra(ReviewActivity.EXTRA_MENU_ID, R.menu.bottom_navigation_menu_course_review) // can be any
-        intent.putExtra(ReviewActivity.EXTRA_GRAPH_ID, R.navigation.nav_graph_course_review)
         intent.putExtra(ReviewActivity.EXTRA_ITEM_REVIEWED_ID, "Fake id")
         intent.putExtra(ReviewActivity.EXTRA_ITEM_REVIEWED, reviewable)
         scenario = ActivityScenario.launch(intent)
-        onView(withId(R.id.reviewBottomNavigationView)).perform(CustomViewActions.navigateTo(R.id.addReviewFragment))
+        ViewPagerAction.swipeNext()
+        Thread.sleep(1000)
+        onView(withId(R.id.addReviewButton)).perform(click())
     }
 
     @After
@@ -52,7 +50,7 @@ class AddReviewFragmentTest {
     }
 
     @Test
-    fun nullGradeNoReset() {
+    fun nullGradeDoesNotWork() {
         val comment = "Good"
         onView(withId(R.id.addReviewComment)).perform(typeText(comment))
         closeSoftKeyboard()
@@ -61,7 +59,7 @@ class AddReviewFragmentTest {
     }
 
     @Test
-    fun nullCommentNoReset() {
+    fun nullCommentDoesNotWork() {
         val title = "Good"
         onView(withId(R.id.reviewRatingBar)).perform(click())
         onView(withId(R.id.addReviewTitle)).perform(typeText(title))
@@ -71,7 +69,7 @@ class AddReviewFragmentTest {
     }
 
     @Test
-    fun nullTitleNoReset() {
+    fun nullTitleDoesNotWork() {
         val comment = "Good"
         onView(withId(R.id.reviewRatingBar)).perform(performSetRating(ReviewRating.GOOD))
         onView(withId(R.id.addReviewComment)).perform(typeText(comment))
@@ -82,7 +80,7 @@ class AddReviewFragmentTest {
 
 
     @Test
-    fun nonNullArgumentsResetsAddReview() {
+    fun nonNullArgumentsAddsReview() {
         val comment = "Good"
         val title = "Good title"
         onView(withId(R.id.reviewRatingBar)).perform(
@@ -95,14 +93,12 @@ class AddReviewFragmentTest {
         onView(withId(R.id.addReviewTitle)).perform(typeText(title))
         closeSoftKeyboard()
         onView(withId(R.id.doneButton)).perform(click())
-        onView(withId(R.id.addReviewComment)).check(matches(withText("")))
-        onView(withId(R.id.addReviewTitle)).check(matches(withText("")))
+        onView(withId(R.id.reviewRecyclerView)).check(matches(isDisplayed()))
     }
 
     @Test
-    fun userNotConnectedNoReset() {
+    fun userNotConnectedDoesNotWork() {
         FakeConnectedUser.instance = FakeConnectedUser.Instance.LOGGED_OUT
-        onView(withId(R.id.reviewBottomNavigationView)).perform(CustomViewActions.navigateTo(R.id.addReviewFragment))
         val comment = "Good"
         val title = "Good title"
         onView(withId(R.id.reviewRatingBar)).perform(
@@ -150,7 +146,6 @@ class AddReviewFragmentTest {
         onView(withId(R.id.addReviewTitle)).perform(typeText(title))
         closeSoftKeyboard()
         onView(withId(R.id.doneButton)).perform(click())
-        onView(withId(R.id.addReviewComment)).check(matches(withText("")))
-        onView(withId(R.id.addReviewTitle)).check(matches(withText("")))
+        onView(withId(R.id.reviewRecyclerView)).check(matches(isDisplayed()))
     }
 }
