+++ app/src/androidTest/java/com/github/sdp/ratemyepfl/fragment/review/ReviewListFragmentTest.kt
@@ -21,7 +21,7 @@ import com.github.sdp.ratemyepfl.model.ImageFile
 import com.github.sdp.ratemyepfl.model.review.Review
 import com.github.sdp.ratemyepfl.model.review.ReviewRating
 import com.github.sdp.ratemyepfl.model.serializer.putExtra
-import com.github.sdp.ratemyepfl.utils.CustomViewActions
+import com.github.sdp.ratemyepfl.utils.CustomViewActions.ViewPagerAction
 import com.github.sdp.ratemyepfl.utils.TestUtils.isExpanded
 import com.github.sdp.ratemyepfl.utils.TestUtils.isHidden
 import com.github.sdp.ratemyepfl.utils.TestUtils.resourceToBitmap
@@ -31,8 +31,11 @@ import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.hamcrest.Matchers.allOf
 import org.hamcrest.Matchers.not
-import org.junit.*
+import org.junit.After
 import org.junit.Assert.assertEquals
+import org.junit.FixMethodOrder
+import org.junit.Rule
+import org.junit.Test
 import org.junit.runners.MethodSorters
 import java.time.LocalDate
 
@@ -42,8 +45,6 @@ class ReviewListFragmentTest {
     lateinit var scenario: ActivityScenario<ReviewActivity>
 
     private val intent = Intent(ApplicationProvider.getApplicationContext(), ReviewActivity::class.java)
-        .putExtra(ReviewActivity.EXTRA_MENU_ID, R.menu.bottom_navigation_menu_course_review)
-        .putExtra(ReviewActivity.EXTRA_GRAPH_ID, R.navigation.nav_graph_course_review)
         .putExtra(ReviewActivity.EXTRA_ITEM_REVIEWED_ID, "Fake id")
         .putExtra(ReviewActivity.EXTRA_ITEM_REVIEWED, FakeCourseRepository.COURSE_LIST.first())
 
@@ -53,7 +54,8 @@ class ReviewListFragmentTest {
     fun launch() {
         FakeConnectedUser.instance = FakeConnectedUser.Instance.FAKE_USER_1
         scenario = ActivityScenario.launch(intent)
-        onView(withId(R.id.reviewBottomNavigationView)).perform(CustomViewActions.navigateTo(R.id.reviewListFragment))
+        ViewPagerAction.swipeNext()
+        Thread.sleep(1000)
     }
 
     @After
@@ -62,8 +64,9 @@ class ReviewListFragmentTest {
     }
 
     private fun refresh(){
-        onView(withId(R.id.reviewBottomNavigationView)).perform(CustomViewActions.navigateTo(R.id.courseReviewInfoFragment))
-        onView(withId(R.id.reviewBottomNavigationView)).perform(CustomViewActions.navigateTo(R.id.reviewListFragment))
+        ViewPagerAction.swipePrevious()
+        ViewPagerAction.swipeNext()
+        Thread.sleep(1000)
     }
 
     /**
