+++ app/src/androidTest/java/com/github/sdp/ratemyepfl/fragment/review/RoomReviewInfoFragmentTest.kt
@@ -17,11 +17,9 @@ import androidx.test.rule.GrantPermissionRule
 import com.github.sdp.ratemyepfl.R
 import com.github.sdp.ratemyepfl.activity.ReviewActivity
 import com.github.sdp.ratemyepfl.database.fakes.FakeClassroomRepository
-import com.github.sdp.ratemyepfl.database.fakes.FakeGradeInfoRepository
 import com.github.sdp.ratemyepfl.database.fakes.FakeReviewsRepository
-import com.github.sdp.ratemyepfl.model.GradeInfo
 import com.github.sdp.ratemyepfl.model.serializer.putExtra
-import com.github.sdp.ratemyepfl.utils.CustomViewActions
+import com.github.sdp.ratemyepfl.utils.CustomViewActions.ViewPagerAction
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.After
@@ -47,8 +45,6 @@ class RoomReviewInfoFragmentTest {
 
     private fun launch() {
         val intent = Intent(ApplicationProvider.getApplicationContext(), ReviewActivity::class.java)
-        intent.putExtra(ReviewActivity.EXTRA_MENU_ID, R.menu.bottom_navigation_menu_room_review)
-        intent.putExtra(ReviewActivity.EXTRA_GRAPH_ID, R.navigation.nav_graph_room_review)
         intent.putExtra(ReviewActivity.EXTRA_ITEM_REVIEWED_ID, "Fake id")
         intent.putExtra(ReviewActivity.EXTRA_ITEM_REVIEWED, FakeClassroomRepository.CLASSROOM_LIST.first())
         scenario = ActivityScenario.launch(intent)
@@ -64,8 +60,9 @@ class RoomReviewInfoFragmentTest {
         intending(toPackage("com.github.sdp.ratemyepfl")).respondWith(result)
         onView(withId(R.id.noiseMeasureButton)).perform(click())
 
-        onView(withId(R.id.reviewBottomNavigationView)).perform(CustomViewActions.navigateTo(R.id.reviewListFragment))
-        onView(withId(R.id.reviewBottomNavigationView)).perform(CustomViewActions.navigateTo(R.id.roomReviewInfoFragment))
+        ViewPagerAction.swipeNext()
+        ViewPagerAction.swipePrevious()
+        Thread.sleep(1000)
 
         val dbText = "Calm (50 dB) - 0 minutes ago"
         onView(withId(R.id.roomNoiseInfoTextView)).check(matches(withText(dbText)))
@@ -90,10 +87,6 @@ class RoomReviewInfoFragmentTest {
         FakeClassroomRepository.roomById = FakeClassroomRepository.ROOM_NO_REVIEW
 
         launch()
-
-        // Refresh
-        onView(withId(R.id.reviewBottomNavigationView)).perform(CustomViewActions.navigateTo(R.id.reviewListFragment))
-        onView(withId(R.id.reviewBottomNavigationView)).perform(CustomViewActions.navigateTo(R.id.roomReviewInfoFragment))
 
         val numReviewText = "(No review submitted)"
         onView(withId(R.id.roomNumReview)).check(matches(withText(numReviewText)))
