+++ app/src/main/java/com/github/sdp/ratemyepfl/fragment/review/ReviewListFragment.kt
@@ -7,6 +7,7 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
+import androidx.navigation.Navigation
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
@@ -18,6 +19,7 @@ import com.github.sdp.ratemyepfl.model.review.Review
 import com.github.sdp.ratemyepfl.model.user.User
 import com.github.sdp.ratemyepfl.viewmodel.ReviewListViewModel
 import com.github.sdp.ratemyepfl.viewmodel.UserViewModel
+import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.snackbar.Snackbar
 import com.sothree.slidinguppanel.SlidingUpPanelLayout
 import dagger.hilt.android.AndroidEntryPoint
@@ -64,15 +66,6 @@ class ReviewListFragment : Fragment(R.layout.fragment_review_list) {
         profilePanel.setFadeOnClickListener {
             profilePanel.panelState = SlidingUpPanelLayout.PanelState.HIDDEN
         }
-        reviewAdapter = ReviewAdapter(viewLifecycleOwner, userViewModel,
-            getListener { r, s -> viewModel.updateUpVotes(r, s) },
-            getListener { r, s -> viewModel.updateDownVotes(r, s) },
-            { rwa -> lifecycleScope.launch {
-                viewModel.removeReview(rwa.review.getId())
-            }}
-        ) { rwa -> displayProfilePanel(rwa.author, rwa.image) }
-
-        recyclerView.adapter = reviewAdapter
 
         recyclerView.addItemDecoration(
             DividerItemDecoration(view.context, DividerItemDecoration.VERTICAL)
@@ -82,6 +75,26 @@ class ReviewListFragment : Fragment(R.layout.fragment_review_list) {
             viewModel.updateReviewsList()
             swipeRefresher.isRefreshing = false
         }
+
+        setUpAdapter()
+
+        view.findViewById<FloatingActionButton>(R.id.addReviewButton).setOnClickListener {
+            Navigation.findNavController(view).navigate(R.id.addReviewFragment)
+        }
+    }
+
+    private fun setUpAdapter() {
+        reviewAdapter = ReviewAdapter(viewLifecycleOwner, userViewModel,
+            getListener { r, s -> viewModel.updateUpVotes(r, s) },
+            getListener { r, s -> viewModel.updateDownVotes(r, s) },
+            { rwa ->
+                lifecycleScope.launch {
+                    viewModel.removeReview(rwa.review.getId())
+                }
+            }
+        ) { rwa -> displayProfilePanel(rwa.author, rwa.image) }
+
+        recyclerView.adapter = reviewAdapter
 
         viewModel.reviews.observe(viewLifecycleOwner) {
             it?.let { reviewAdapter.submitList(it) }
