+++ app/src/main/java/com/github/sdp/ratemyepfl/model/items/EventItem.kt
@@ -1,9 +1,8 @@
 package com.github.sdp.ratemyepfl.model.items
 
-import com.github.sdp.ratemyepfl.R
 import com.google.android.gms.maps.model.BitmapDescriptor
 import com.google.android.gms.maps.model.LatLng
 
 class EventItem(event: Event, photo: Int, icon: BitmapDescriptor?) :
-    MapItem(LatLng(event.lat, event.long), event.getId(), photo, icon) {
+    MapItem(LatLng(event.lat, event.long), event, event.getId(), photo, icon) {
 }
