+++ app/src/main/res/layout/fragment_review_list.xml
@@ -10,18 +10,37 @@
 
     <!-- MAIN CONTENT: REVIEW LIST -->
 
-    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-        android:id="@+id/reviewSwipeRefresh"
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/reviewRecyclerView"
+        <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+            android:id="@+id/reviewSwipeRefresh"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:scrollbars="vertical"
-            app:layoutManager="LinearLayoutManager" />
-    </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
+            android:layout_height="match_parent">
+
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/reviewRecyclerView"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:scrollbars="vertical"
+                app:layoutManager="LinearLayoutManager" />
+        </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
+
+        <com.google.android.material.floatingactionbutton.FloatingActionButton
+            android:id="@+id/addReviewButton"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginEnd="15dp"
+            android:layout_marginBottom="15dp"
+            android:backgroundTint="@color/white"
+            android:contentDescription="@string/add_review_button"
+            android:elevation="50dp"
+            android:src="@drawable/ic_add_circle_white_24dp"
+            app:borderWidth="0dp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
     <!-- DRAGGABLE VIEW: SELECTED AUTHOR PROFILE -->
 
@@ -41,8 +60,8 @@
                 android:id="@+id/author_panel_username"
                 android:layout_width="180dp"
                 android:layout_height="wrap_content"
-                android:layout_marginStart="145dp"
                 android:layout_alignParentBottom="true"
+                android:layout_marginStart="145dp"
                 android:layout_marginBottom="20dp"
                 android:background="@android:color/transparent"
                 android:hint="@string/anonymous"
