+++ app/src/main/res/navigation/reviewable.xml
@@ -8,6 +8,6 @@
     <fragment
         android:id="@+id/reviewableFragment"
         android:name="com.github.sdp.ratemyepfl.fragment.navigation.ReviewableFragment"
-        android:label="@string/reviews_nav_title"
+        android:label="@string/reviewable_nav_title"
         tools:layout="@layout/fragment_review" />
 </navigation>
