+++ app/src/main/res/values/strings.xml
@@ -31,7 +31,7 @@
     <string name="title_edit_text">Give a title to your review</string>
     <string name="maps_api_key" tools:ignore="TypographyDashes,Typos">AIzaSyBmyu-adSXBHO7qfyQ4SoAyPt6YO--At1U</string>
     <string name="home_nav_title">Home</string>
-    <string name="reviews_nav_title">Reviewables</string>
+    <string name="reviewable_nav_title">Reviewables</string>
     <string name="events_nav_item">Events</string>
     <string name="map_nav_title">Map</string>
     <string name="home_page_title">HOME PAGE</string>
@@ -131,4 +131,5 @@
     <string name="title_class_creation">Add a class</string>
     <string name="day_class_creation">Select the day and time</string>
     <string name="not_logged_in_text">You are not logged in</string>
+    <string name="add_review_button">Add Review Button</string>
 </resources>
